workflow: slim transition ownership
