+++ packages/core/src/sprints/workflows/lanes.py
@@ -7,6 +7,10 @@ submodule directly.
 from __future__ import annotations
 
 from sprints.workflows.intake import claim_new_lanes
+from sprints.workflows.actor_outputs import (
+    apply_actor_output_status,
+    record_actor_output,
+)
 from sprints.workflows.lane_state import (
     active_lanes,
     lane_actor_runtime_session,
@@ -30,14 +34,12 @@ from sprints.workflows.status import (
 from sprints.workflows.transitions import (
     actor_concurrency_usage,
     advance_lane,
-    apply_actor_output_status,
     complete_lane,
     decision_ready_lanes,
     guard_actor_dispatch,
     lane_for_decision,
     record_action_result,
     record_actor_dispatch_planned,
-    record_actor_output,
     record_actor_runtime_progress,
     record_actor_runtime_result,
     record_actor_runtime_start,
