+++ packages/core/src/sprints/workflows/teardown.py
@@ -3,18 +3,12 @@
 from __future__ import annotations
 
 import re
-import time
 from collections.abc import Callable
 from dataclasses import dataclass
-from datetime import datetime
-from pathlib import Path
 from typing import Any
 
 from sprints.engine import (
-    EngineStore,
-    RetryPolicy,
     pending_retry_projection,
-    retry_is_due,
     retry_record,
 )
 from sprints.trackers import build_code_host_client, build_tracker_client
@@ -29,10 +23,24 @@ from sprints.workflows.effects import (
 )
 from sprints.workflows.completion import done_release_verified
 from sprints.workflows.orchestrator import OrchestratorDecision
-from sprints.core.paths import runtime_paths
-
-_TERMINAL_LANE_STATUSES = {"complete", "released"}
-
+from sprints.workflows.lane_state import (
+    code_host_config as _code_host_config,
+    configured_texts as _configured_texts,
+    engine_store as _engine_store,
+    issue_labels as _issue_labels,
+    lane_is_terminal as _lane_is_terminal,
+    lane_list as _lane_list,
+    lane_mapping as _lane_mapping,
+    lane_run_id as _lane_run_id,
+    lane_stage as _lane_stage,
+    lane_transition_side as _lane_transition_side,
+    now_iso as _now_iso,
+    repository_path as _repository_path,
+    retry_engine_entry as _retry_engine_entry,
+    retry_policy as _retry_policy,
+    tracker_config as _tracker_config,
+)
+from sprints.workflows.retries import lane_retry_is_due as _lane_retry_is_due
 
 @dataclass(frozen=True)
 class TeardownOps:
@@ -853,55 +861,6 @@ def _completion_auto_merge_config(config: WorkflowConfig) -> dict[str, Any]:
     }
 
 
-def _retry_policy(config: WorkflowConfig) -> RetryPolicy:
-    cfg = _retry_config(config)
-    return RetryPolicy(
-        max_attempts=cfg["max_attempts"],
-        initial_delay_seconds=cfg["initial_delay_seconds"],
-        backoff_multiplier=cfg["backoff_multiplier"],
-        max_delay_seconds=cfg["max_delay_seconds"],
-    )
-
-
-def _retry_config(config: WorkflowConfig) -> dict[str, Any]:
-    raw = config.raw.get("retry")
-    cfg = raw if isinstance(raw, dict) else {}
-    return {
-        "max_attempts": _positive_int(cfg, "max-attempts", "max_attempts", default=3),
-        "initial_delay_seconds": _nonnegative_int(
-            cfg,
-            "initial-delay-seconds",
-            "initial_delay_seconds",
-            default=0,
-        ),
-        "backoff_multiplier": _positive_float(
-            cfg,
-            "backoff-multiplier",
-            "backoff_multiplier",
-            default=2.0,
-        ),
-        "max_delay_seconds": _nonnegative_int(
-            cfg,
-            "max-delay-seconds",
-            "max_delay_seconds",
-            default=300,
-        ),
-    }
-
-
-def _retry_engine_entry(lane: dict[str, Any]) -> dict[str, Any]:
-    issue = lane.get("issue") if isinstance(lane.get("issue"), dict) else {}
-    return {
-        **_scheduler_entry(lane),
-        "issue_id": lane.get("lane_id"),
-        "identifier": issue.get("identifier") or lane.get("lane_id"),
-        "error": "retry queued",
-        "current_attempt": int(lane.get("attempt") or 0),
-        "delay_type": "workflow-retry",
-        "run_id": _lane_run_id(lane),
-    }
-
-
 def _retry_event_retry(retry: dict[str, Any]) -> dict[str, Any]:
     return {
         "status": retry.get("status"),
@@ -918,146 +877,6 @@ def _retry_event_retry(retry: dict[str, Any]) -> dict[str, Any]:
         "queued_at": retry.get("queued_at"),
     }
 
-
-def _scheduler_entry(lane: dict[str, Any]) -> dict[str, Any]:
-    issue = lane.get("issue") if isinstance(lane.get("issue"), dict) else {}
-    session = (
-        lane.get("runtime_session")
-        if isinstance(lane.get("runtime_session"), dict)
-        else {}
-    )
-    return {
-        "issue_id": lane.get("lane_id"),
-        "identifier": issue.get("identifier") or lane.get("lane_id"),
-        "state": lane.get("status"),
-        "title": issue.get("title"),
-        "url": issue.get("url"),
-        "worker_id": lane.get("actor"),
-        "attempt": int(lane.get("attempt") or 0),
-        "worker_status": lane.get("status"),
-        "started_at_epoch": _iso_to_epoch(
-            str(session.get("started_at") or lane.get("last_progress_at") or ""),
-            default=time.time(),
-        ),
-        "heartbeat_at_epoch": _iso_to_epoch(
-            str(session.get("updated_at") or lane.get("last_progress_at") or ""),
-            default=time.time(),
-        ),
-        "thread_id": lane.get("thread_id") or session.get("thread_id"),
-        "turn_id": lane.get("turn_id") or session.get("turn_id"),
-        "session_name": session.get("session_name"),
-        "runtime_name": session.get("runtime_name"),
-        "runtime_kind": session.get("runtime_kind"),
-        "session_id": session.get("session_id"),
-        "status": session.get("status") or lane.get("status"),
-        "run_id": session.get("run_id"),
-        "actor": session.get("actor") or lane.get("actor"),
-        "stage": session.get("stage") or lane.get("stage"),
-        "branch": lane.get("branch"),
-        "pull_request": lane.get("pull_request"),
-    }
-
-
-def _lane_retry_is_due(lane: dict[str, Any], *, now_epoch: float | None = None) -> bool:
-    pending = (
-        lane.get("pending_retry") if isinstance(lane.get("pending_retry"), dict) else {}
-    )
-    return retry_is_due(pending, now_epoch=now_epoch)
-
-
-def _lane_transition_side(lane: dict[str, Any]) -> dict[str, Any]:
-    claim = lane.get("claim") if isinstance(lane.get("claim"), dict) else {}
-    return {
-        "status": lane.get("status"),
-        "stage": lane.get("stage"),
-        "actor": lane.get("actor"),
-        "attempt": lane.get("attempt"),
-        "claim_state": claim.get("state"),
-        "pending_retry": lane.get("pending_retry"),
-        "operator_attention": lane.get("operator_attention"),
-    }
-
-
-def _engine_store(config: WorkflowConfig) -> EngineStore:
-    return EngineStore(
-        db_path=runtime_paths(config.workflow_root)["db_path"],
-        workflow=config.workflow_name,
-    )
-
-
-def _tracker_config(config: WorkflowConfig) -> dict[str, Any]:
-    raw = config.raw.get("tracker")
-    return raw if isinstance(raw, dict) else {}
-
-
-def _code_host_config(config: WorkflowConfig) -> dict[str, Any]:
-    raw = config.raw.get("code-host")
-    return raw if isinstance(raw, dict) else {}
-
-
-def _repository_path(config: WorkflowConfig) -> Path | None:
-    raw = config.raw.get("repository")
-    if not isinstance(raw, dict):
-        return None
-    value = str(raw.get("local-path") or raw.get("local_path") or "").strip()
-    if not value:
-        return None
-    path = Path(value).expanduser()
-    return path if path.is_absolute() else (config.workflow_root / path).resolve()
-
-
-def _lane_is_terminal(lane: dict[str, Any]) -> bool:
-    return str(lane.get("status") or "").strip() in _TERMINAL_LANE_STATUSES
-
-
-def _lane_stage(lane: dict[str, Any]) -> str:
-    return str(lane.get("stage") or "").strip()
-
-
-def _lane_mapping(lane: dict[str, Any], key: str) -> dict[str, Any]:
-    value = lane.get(key)
-    if isinstance(value, dict):
-        return value
-    lane[key] = {}
-    return lane[key]
-
-
-def _lane_list(lane: dict[str, Any], key: str) -> list[Any]:
-    value = lane.get(key)
-    if isinstance(value, list):
-        return value
-    lane[key] = []
-    return lane[key]
-
-
-def _lane_run_id(lane: dict[str, Any]) -> str | None:
-    session = (
-        lane.get("runtime_session")
-        if isinstance(lane.get("runtime_session"), dict)
-        else {}
-    )
-    value = session.get("run_id")
-    text = str(value or "").strip()
-    return text or None
-
-
-def _issue_labels(issue: dict[str, Any]) -> set[str]:
-    labels: set[str] = set()
-    for label in issue.get("labels") or []:
-        text = str(label.get("name") if isinstance(label, dict) else label).strip()
-        if text:
-            labels.add(text.lower())
-    return labels
-
-
-def _configured_texts(config: dict[str, Any], *keys: str) -> list[str]:
-    for key in keys:
-        value = config.get(key)
-        if isinstance(value, list):
-            return [str(item).strip().lower() for item in value if str(item).strip()]
-    return []
-
-
 def _configured_bool(config: dict[str, Any], *keys: str, default: bool) -> bool:
     for key in keys:
         value = config.get(key)
@@ -1071,56 +890,3 @@ def _configured_bool(config: dict[str, Any], *keys: str, default: bool) -> bool:
         if text in {"0", "false", "no", "off"}:
             return False
     return default
-
-
-def _positive_int(config: dict[str, Any], *keys: str, default: int) -> int:
-    for key in keys:
-        parsed = _positive_int_value(config.get(key))
-        if parsed is not None:
-            return parsed
-    return default
-
-
-def _positive_int_value(value: Any) -> int | None:
-    if value in (None, ""):
-        return None
-    try:
-        return max(int(value), 1)
-    except (TypeError, ValueError):
-        return None
-
-
-def _nonnegative_int(config: dict[str, Any], *keys: str, default: int) -> int:
-    for key in keys:
-        value = config.get(key)
-        if value not in (None, ""):
-            try:
-                return max(int(value), 0)
-            except (TypeError, ValueError):
-                return default
-    return default
-
-
-def _positive_float(config: dict[str, Any], *keys: str, default: float) -> float:
-    for key in keys:
-        value = config.get(key)
-        if value not in (None, ""):
-            try:
-                return max(float(value), 1.0)
-            except (TypeError, ValueError):
-                return default
-    return default
-
-
-def _iso_to_epoch(value: str, *, default: float) -> float:
-    text = str(value or "").strip()
-    if not text:
-        return default
-    try:
-        return datetime.fromisoformat(text.replace("Z", "+00:00")).timestamp()
-    except ValueError:
-        return default
-
-
-def _now_iso() -> str:
-    return time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime())
